Revert inadvertent merge before review for LUCENE-9302, SOLR-14381

+++ lucene/grouping/src/java/org/apache/lucene/search/grouping/BlockGroupingCollector.java
@@ -353,8 +353,8 @@ public class BlockGroupingCollector extends SimpleCollector {
 
     return new TopGroups<>(new TopGroups<>(groupSort.getSort(),
                                        withinGroupSort.getSort(),
-                                      (long) totalHitCount, (long) totalGroupedHitCount, groups, maxScore),
-                          (long) totalGroupCount);
+                                       totalHitCount, totalGroupedHitCount, groups, maxScore),
+                         totalGroupCount);
   }
 
   @Override
